datamux/requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements.dev.txt --strip-extras pyproject.toml
#
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
avro==1.11.3
    # via streaminghub-datamux (pyproject.toml)
black==23.12.1
    # via streaminghub-datamux (pyproject.toml)
build==1.0.3
    # via
    #   pip-tools
    #   streaminghub-datamux (pyproject.toml)
bumpver==2023.1129
    # via streaminghub-datamux (pyproject.toml)
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   bumpver
    #   pip-tools
    #   streaminghub-datamux (pyproject.toml)
colorama==0.4.6
    # via bumpver
contourpy==1.2.0
    # via matplotlib
cryptography==41.0.7
    # via secretstorage
cycler==0.12.1
    # via matplotlib
docutils==0.20.1
    # via readme-renderer
exceptiongroup==1.2.0
    # via pytest
fonttools==4.47.0
    # via matplotlib
h5py==3.10.0
    # via
    #   streaminghub-datamux (pyproject.toml)
    #   streaminghub-pydfds
idna==3.6
    # via requests
importlib-metadata==7.0.1
    # via
    #   keyring
    #   twine
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via streaminghub-datamux (pyproject.toml)
jaraco-classes==3.3.0
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jsonschema==4.20.0
    # via
    #   streaminghub-datamux (pyproject.toml)
    #   streaminghub-pydfds
jsonschema-specifications==2023.12.1
    # via jsonschema
keyring==24.3.0
    # via twine
kiwisolver==1.4.5
    # via matplotlib
lexid==2021.1006
    # via bumpver
looseversion==1.3.0
    # via bumpver
lxml==5.0.0
    # via streaminghub-datamux (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
matplotlib==3.8.2
    # via streaminghub-datamux (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.1.0
    # via jaraco-classes
msgpack==1.0.7
    # via streaminghub-datamux (pyproject.toml)
mypy-extensions==1.0.0
    # via black
nh3==0.2.15
    # via readme-renderer
numpy==1.26.3
    # via
    #   contourpy
    #   h5py
    #   matplotlib
    #   pandas
    #   pyarrow
    #   streaminghub-datamux (pyproject.toml)
    #   streaminghub-pydfds
packaging==23.2
    # via
    #   black
    #   build
    #   matplotlib
    #   pytest
pandas==2.1.4
    # via
    #   streaminghub-datamux (pyproject.toml)
    #   streaminghub-pydfds
parse==1.20.0
    # via streaminghub-pydfds
pathspec==0.12.1
    # via black
pillow==10.2.0
    # via
    #   matplotlib
    #   streaminghub-datamux (pyproject.toml)
pip-tools==7.3.0
    # via streaminghub-datamux (pyproject.toml)
pkginfo==1.9.6
    # via twine
platformdirs==4.1.0
    # via black
pluggy==1.3.0
    # via pytest
pyarrow==14.0.2
    # via
    #   streaminghub-datamux (pyproject.toml)
    #   streaminghub-pydfds
pycparser==2.21
    # via cffi
pydantic==2.5.3
    # via streaminghub-pydfds
pydantic-core==2.14.6
    # via pydantic
pygments==2.17.2
    # via
    #   readme-renderer
    #   rich
pylsl==1.16.2
    # via
    #   streaminghub-datamux (pyproject.toml)
    #   streaminghub-pydfds
pyparsing==3.1.1
    # via matplotlib
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via streaminghub-datamux (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via streaminghub-datamux (pyproject.toml)
pytz==2023.3.post1
    # via pandas
pyzmq==25.1.2
    # via streaminghub-datamux (pyproject.toml)
readme-renderer==42.0
    # via twine
referencing==0.32.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.7.0
    # via
    #   streaminghub-datamux (pyproject.toml)
    #   twine
rpds-py==0.16.2
    # via
    #   jsonschema
    #   referencing
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via python-dateutil
streaminghub-pydfds==0.1.32
    # via streaminghub-datamux (pyproject.toml)
toml==0.10.2
    # via bumpver
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
tqdm==4.66.1
    # via streaminghub-datamux (pyproject.toml)
twine==4.0.2
    # via streaminghub-datamux (pyproject.toml)
typing-extensions==4.9.0
    # via
    #   black
    #   pydantic
    #   pydantic-core
tzdata==2023.4
    # via pandas
urllib3==2.1.0
    # via
    #   requests
    #   twine
websockets==12.0
    # via streaminghub-datamux (pyproject.toml)
wheel==0.42.0
    # via pip-tools
xmltodict==0.13.0
    # via streaminghub-datamux (pyproject.toml)
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
